% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_api_pred.R
\name{api_pred}
\alias{api_pred}
\title{api_pred}
\usage{
api_pred(endpoint, json)
}
\arguments{
\item{endpoint}{API endpoint}

\item{json}{JSON list of dictionaries with the following compulsory keys:
`comment_id` (string) and `comment_text` (string). For example, 
`[{'comment_id': '1', 'comment_text': 'Thank you'}, {'comment_id': '2', 'comment_text': 'Food was cold'}]`}
}
\value{
a dataframe
}
\description{
A function to call the `pxtextmining` API
}
