% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_api_pred.R
\name{api_pred_url}
\alias{api_pred_url}
\title{Get the prediction URL or the data from the `pxtextmining API`}
\usage{
api_pred_url(data, api_key = Sys.getenv("API_Key"))
}
\arguments{
\item{data}{Dataframe with column `comment_id`, `comment_text` and `question_type`}

\item{api_key}{api key to access the api}
}
\value{
a dataframe or a url to get the data
}
\description{
A function to call the `pxtextmining` API
}
