% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment_functions.R
\name{calc_sentiment}
\alias{calc_sentiment}
\title{Produce sentiment dataframe for tidying}
\usage{
calc_sentiment(data, sentiment_names)
}
\arguments{
\item{data}{dataframe- raw from the database}

\item{sentiment_names}{all of the nrc sentiments in a vector of strings}
}
\value{
dataframe with sentiment columns
}
\description{
Take the raw dataframe and mark it up for sentiment analysis
}
