% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_table.R
\name{calculate_table}
\alias{calculate_table}
\title{calculate percentage tables for comment themes}
\usage{
calculate_table(table_data, count_column, comment_type = NULL)
}
\arguments{
\item{table_data}{A dataframe}

\item{count_column}{String with the name of the column to count in the joined table}

\item{comment_type}{String with type of comment}
}
\value{
a dataframe with category name, n, and %
}
\description{
calculate percentage tables for comment themes
}
