% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_upload.R
\name{clean_dataframe}
\alias{clean_dataframe}
\title{Clean an uploaded dataset from a user of the dashboard}
\usage{
clean_dataframe(data, text_cols)
}
\arguments{
\item{data}{a dataframe of uploaded patient experience data}

\item{text_cols}{a vector of strings with the name of the text columns}
}
\value{
dataframe with cleaned text
}
\description{
uploaded data will often have garbage in the comments 
box such as "NULL", "NA", "N/A", etc. Clean these comments before they go 
to the pipeline
}
