% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demographic_graphs.R
\name{demographic_distribution}
\alias{demographic_distribution}
\title{Draw the distribution of demographics in the sample}
\usage{
demographic_distribution(pass_data, variable)
}
\arguments{
\item{pass_data}{A dataframe, filtered to unique individuals}

\item{variable}{String. Name of variable- "age", "gender", "ethnicity"}
}
\value{
a ggplot2 graph
}
\description{
Draw the distribution of demographics in the sample
}
