% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\name{drop_na_for_col}
\alias{drop_na_for_col}
\title{Find rows containing missing values in all specified 
column while keeping rows where any contains values}
\usage{
drop_na_for_col(df, vars, negate = TRUE)
}
\arguments{
\item{df}{A data frame}

\item{vars}{list of strings containing the columns}

\item{negate}{logical, whether to return the rows where the NA check is true or otherwise
default is `TRUE`}
}
\value{
data frame
}
\description{
Find rows containing missing values in all specified 
column while keeping rows where any contains values
}
