% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_text.R
\name{filter_df}
\alias{filter_df}
\title{Filter a dataframe based on specified comment type}
\usage{
filter_df(text_data, comment_type_filter)
}
\arguments{
\item{text_data}{the dataframe from the database}

\item{comment_type_filter}{comment type value to be used to filter}
}
\value{
list of strings
}
\description{
Filter a dataframe based on specified comment type
}
