% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_api_pred.R
\name{get_pred_from_url}
\alias{get_pred_from_url}
\title{Get the prediction from the URL returned from the pxtextmining API}
\usage{
get_pred_from_url(api_url)
}
\arguments{
\item{api_url}{the url return from `get_api_pred_url()` where you can get the prediction data when its ready}
}
\value{
a dataframe or string depending on whether the API is done prediction or not
}
\description{
Get the prediction from the URL returned from the pxtextmining API
}
