% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-pipe.R
\name{html_decoder}
\alias{html_decoder}
\title{html decoder fo single string object}
\usage{
html_decoder(str)
}
\arguments{
\item{str}{a string to be decoded}
}
\value{
decoded string
}
\description{
html decoder fo single string object
}
\examples{
html_decoder("&amp; &gt;")
}
