% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_text.R
\name{input_sanitizer}
\alias{input_sanitizer}
\title{Takes input strings and splits them/sanitizes them}
\usage{
input_sanitizer(filter_text)
}
\arguments{
\item{filter_text}{comma separated string with search terms in}
}
\value{
string as a vector
}
\description{
split list of strings by commas and remove all non-alphanumeric characters form them
}
