% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment_functions.R
\name{make_sentiment_table}
\alias{make_sentiment_table}
\title{Produce a table of comments with different sentiments
for example to be drawn with {reactable}}
\usage{
make_sentiment_table(data, sentiment_names)
}
\arguments{
\item{data}{dataframe you almost certainly made with 
\code{\link{tidy_sentiment_txt}}}

\item{sentiment_names}{all of the nrc sentiments in a vector of strings}
}
\value{
a table suitable for {reactable}
}
\description{
Produce a table of comments with different sentiments
for example to be drawn with {reactable}
}
