% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_text.R
\name{matched_comments}
\alias{matched_comments}
\title{Find  rows with matching string(s)}
\usage{
matched_comments(split_comments, search_fn, search_strings)
}
\arguments{
\item{split_comments}{list of tokenized words}

\item{search_fn}{type of search ('and', 'or')}

\item{search_strings}{list of strings with search terms in it}
}
\description{
Find  rows with matching string(s)
}
