% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_overlap_plot_helpers.R
\name{multi_to_single_label}
\alias{multi_to_single_label}
\title{multilabeled_to_singlelabel}
\usage{
multi_to_single_label(mt_data, column_name, n_labels = 10)
}
\arguments{
\item{mt_data}{A dataframe with a single column for the comma separated labels for each row}

\item{column_name}{The name of the column holding the comma separated labels}

\item{n_labels}{maximum number of labels assigned to any row}
}
\value{
A dataframe
}
\description{
Convert a multi-labeled data to a single-label data with a category per comment per row
}
