% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc_functions.R
\name{plot_fft_spc}
\alias{plot_fft_spc}
\title{Plot data}
\usage{
plot_fft_spc(data)
}
\arguments{
\item{data}{dataframe, that you probably made with the split_data_spc function}
}
\value{
SPC plot
}
\description{
Plot data
}
