% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sentiment_functions.R
\name{plot_sentiment}
\alias{plot_sentiment}
\title{Plot sentiment graph over time, faceting by category and/ or location}
\usage{
plot_sentiment(
  data,
  sentiment_names,
  select_sentiment,
  select_fill_type,
  select_facet
)
}
\arguments{
\item{data}{reactive and TIDY sentiment object}

\item{sentiment_names}{all of the nrc sentiments in a vector of strings}

\item{select_sentiment}{value of selected sentiments (from input$)}

\item{select_fill_type}{string "fill" or "stack" corresponding to 
arguments to geom_histogram()}

\item{select_facet}{integer 1, 2, 3, corresponding to facet by
category, location, category & location respectively}
}
\description{
Plot sentiment graph over time, faceting by category and/ or location
}
