% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{previous_quarter}
\alias{previous_quarter}
\title{find the previous quarter}
\usage{
previous_quarter(date)
}
\arguments{
\item{date}{date, probably \code{\link{Sys.Date()}}}
}
\value{
vector of two dates, one at the beginning and one at the end of the
quarter previous to the date
}
\description{
find the previous quarter
}
