% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_text.R
\name{split_comments}
\alias{split_comments}
\title{Takes a list of comments and splits them into tokens (words)}
\usage{
split_comments(comments)
}
\arguments{
\item{comments}{list of strings}
}
\value{
list of tokenized words
}
\description{
Takes a list of comments and splits them into tokens (words)
}
