% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spc_functions.R
\name{split_data_spc}
\alias{split_data_spc}
\title{Get data ready for SPC plotting}
\usage{
split_data_spc(data, variable = "service", chunks)
}
\arguments{
\item{data}{dataframe containing "date" column and an FFT column}

\item{variable}{string, indicating the name of the FFT column}

\item{chunks}{either "monthly", which divides the data in months 
or the number of chunks to divide the data into (each chunk will have the 
same number of rows)}

\item{return}{dataframe suitable for plotting with qicharts2}
}
\value{

}
\description{
TODO, ADD WHAT IS IN HERE AN WHAT ISNT
}
