% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{tidy_all_trusts}
\alias{tidy_all_trusts}
\title{Tidy patient experience data}
\usage{
tidy_all_trusts(data)
}
\arguments{
\item{data}{dataframe or SQL object, that you can make with get_px_exp()}
}
\description{
Tidy patient experience data
}
\section{Last updated by}{

Chris Beeley
}

\section{Last updated date}{

2021-04-25
}

