% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{tidy_px_exp}
\alias{tidy_px_exp}
\title{Tidy patient experience data}
\usage{
tidy_px_exp(data, conn, trust_id = "trust_a")
}
\arguments{
\item{data}{dataframe or SQL object, that you can make with get_px_exp()}

\item{conn}{connection, that you can make with connect_mysql()- by default
this will be done automatically}

\item{trust_id}{string. Which trust are you tidying data for (NHC is trust_a)?}
}
\value{

}
\description{
Tidy patient experience data
}
\section{Last updated by}{

Chris Beeley
}

\section{Last updated date}{

2021-04-25
}

