% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils_helpers.R
\name{tidy_sentiment_txt}
\alias{tidy_sentiment_txt}
\title{Tidy sentiment data for further analysis and visualisation}
\usage{
tidy_sentiment_txt(data)
}
\arguments{
\item{data}{Dataframe with sentiment analysis results}
}
\value{

}
\description{
This function currently has no arguments, but will become more flexible with time
}
