% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{verbatim_summary}
\alias{verbatim_summary}
\title{produce a list of comments concatenated with the location after the comment}
\usage{
verbatim_summary(return_data, category_selection)
}
\arguments{
\item{return_data}{dataframe - clean version of the data feed into the module}

\item{category_selection}{list of category to filter the data by}
}
\value{
Vector list of strings
}
\description{
produce a list of comments concatenated with the location after the comment
}
