% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/golem_utils_server.R
\name{add_theme_nhs}
\alias{add_theme_nhs}
\title{Add NHS theme to ggplot chart}
\usage{
add_theme_nhs()
}
\description{
This adds a theme to your chart by tweaking
[NHSRtheme](https://github.com/nhs-r-community/NHSRtheme/blob/main/R/theme_nhs.R)
}
