% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_api_pred.R
\name{api_pred}
\alias{api_pred}
\title{api_pred}
\usage{
api_pred(json)
}
\arguments{
\item{json}{JSON list of dictionaries with the following compulsory keys:
`comment_id` (string), `comment_text` (string) and `question_type` (any of ("what_good", "could_improve", "nonspecific"). For example, 
`[{'comment_id': '1', 'comment_text': 'Thank you', 'question_type': 'what_good'}, {'comment_id': '2', 'comment_text': 'Food was cold', 'question_type': 'could_improve'}]`}
}
\value{
a dataframe
}
\description{
A function to call the `pxtextmining` API
}
