% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_table.R
\name{calculate_table}
\alias{calculate_table}
\title{calculate percentage tables for comment themes}
\usage{
calculate_table(
  table_data,
  count_column,
  comment_type = c("imp", "best"),
  click_column = NULL
)
}
\arguments{
\item{table_data}{A dataframe}

\item{count_column}{String with the name of the column to count in the joined table}

\item{comment_type}{String with type of comment- improve/ best}

\item{click_column}{A string with the name of the category that has been clicked, if any}
}
\value{
a dataframe with category name, n, and %
}
\description{
calculate percentage tables for comment themes
}
