% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_upload.R
\name{clean_dataframe}
\alias{clean_dataframe}
\title{Clean the comment column of uploaded dataset from a user of the dashboard}
\usage{
clean_dataframe(data, comment_column)
}
\arguments{
\item{data}{a dataframe of uploaded patient experience data}

\item{comment_column}{a string for the name of the column for the comment text}
}
\value{
dataframe with cleaned comment_column
}
\description{
uploaded data will often have garbage in the comments
box such as "NULL", "NA", "N/A", etc. Clean these comments and drop all
rows without comment before they go to the pipeline
}
