% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_data.R
\name{tidy_all_trusts}
\alias{tidy_all_trusts}
\title{Tidy patient experience data}
\usage{
tidy_all_trusts(data)
}
\arguments{
\item{data}{dataframe or SQL object with comment column, comment_txt}
}
\description{
Tidy patient experience data
}
