% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fct_api_pred.R
\name{track_api_job}
\alias{track_api_job}
\title{Track the API job table. If prediction is done, it writes it to the main table and delete the job from the api job table}
\usage{
track_api_job(job, conn, write_db = TRUE)
}
\arguments{
\item{job}{an instance of the api job table}

\item{conn}{database connection}

\item{write_db}{logical should the prediction data be written to the database or returned as a dataframe?}
}
\value{
dataframe (if `write_db` is FALSE)
}
\description{
Track the API job table. If prediction is done, it writes it to the main table and delete the job from the api job table
}
