% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tidy_upload.R
\name{upload_data}
\alias{upload_data}
\title{Tidy data upload from spreadsheet}
\usage{
upload_data(data, conn, trust_id)
}
\arguments{
\item{data}{dataframe, loaded within Shiny application}

\item{conn}{connection, from existing {pool}}

\item{trust_id}{string. Which trust are you tidying data for?}
}
\value{
boolean, indicating success or failure in upload
}
\description{
Tidy data upload from spreadsheet
}
\section{Last updated by}{

Chris Beeley
}

\section{Last updated date}{

2021-04-25
}

