% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report_functions.R
\name{verbatim_summary}
\alias{verbatim_summary}
\title{produce a summary of the comments from comment 1 OR 2 and paste the 
location after the comment}
\usage{
verbatim_summary(data, comment_selection)
}
\arguments{
\item{data}{dataframe- the main data fed in to the module}

\item{comment_selection}{string. comment_1 or comment_2}
}
\value{
string. A massive string with line returns suitable for inclusion
in RMarkdown report
}
\description{
produce a summary of the comments from comment 1 OR 2 and paste the 
location after the comment
}
